+++ src/components/pages/Homepage/Homepage.js
@@ -273,7 +273,7 @@ const HomeSExpertsBlock = lazy(() => import("../../organisms/SectionContents/exp
 const HomeSLabBlock = lazy(() => import("../../organisms/SectionContents/lab.js"));
 const HomeSMethodBlock = lazy(() => import("../../organisms/SectionContents/method.js"));
 const HomeSServicesBlock = lazy(() => import("../../organisms/SectionContents/services.js"));
-const HomeSReviewsBlock = lazy(() => import("../../organisms/SectionContents/test.jsx"));
+const HomeSReviewsBlock = lazy(() => import("../../organisms/SectionContents/reviews.jsx"));
 const HomeSPricingBlock = lazy(() => import("../../organisms/SectionContents/pricing.js"));
 const HomeSAboutBlock = lazy(() => import("../../organisms/SectionContents/about.js"));
 const HomeSGalleryBlock = lazy(() => import("../../organisms/SectionContents/gallery"));
